refactor: made separate folder for getVideos
